Updated the top level CMakeLists.txt to include warnings for implicit type conversions and corrected instances in SerialPort and unit tests where implicit conversions were taking place.

+++ src/SerialPort.h
@@ -274,9 +274,9 @@ namespace LibSerial
          * @param numberOfBytes The number of bytes to read before returning.
          * @param msTimeout The timeout period in milliseconds.
          */
-        void Read(char&              charBuffer,
-                  const unsigned int numberOfBytes = 0,
-                  const unsigned int msTimeout  = 0);
+        void Read(char&         charBuffer,
+                  const ssize_t numberOfBytes = 0,
+                  const ssize_t msTimeout  = 0);
 
         /**
          * @brief Reads the specified number of bytes from the serial port.
@@ -290,9 +290,9 @@ namespace LibSerial
          * @param numberOfBytes The number of bytes to read before returning.
          * @param msTimeout The timeout period in milliseconds.
          */
-        void Read(unsigned char&     charBuffer,
-                  const unsigned int numberOfBytes = 0,
-                  const unsigned int msTimeout  = 0);
+        void Read(unsigned char& charBuffer,
+                  const ssize_t  numberOfBytes = 0,
+                  const ssize_t  msTimeout  = 0);
 
         /**
          * @brief Reads the specified number of bytes from the serial port.
@@ -307,9 +307,9 @@ namespace LibSerial
          * @param numberOfBytes The number of bytes to read before returning.
          * @param msTimeout The timeout period in milliseconds.
          */
-        void Read(DataBuffer&        dataBuffer,
-                  const unsigned int numberOfBytes = 0,
-                  const unsigned int msTimeout  = 0);
+        void Read(DataBuffer&   dataBuffer,
+                  const ssize_t numberOfBytes = 0,
+                  const ssize_t msTimeout  = 0);
 
         /**
          * @brief Reads the specified number of bytes from the serial port.
@@ -324,9 +324,9 @@ namespace LibSerial
          * @param numberOfBytes The number of bytes to read before returning.
          * @param msTimeout The timeout period in milliseconds.
          */
-        void Read(std::string&       dataString,
-                  const unsigned int numberOfBytes = 0,
-                  const unsigned int msTimeout  = 0);
+        void Read(std::string&  dataString,
+                  const ssize_t numberOfBytes = 0,
+                  const ssize_t msTimeout  = 0);
 
         /**
          * @brief Reads a single byte from the serial port.
@@ -337,8 +337,8 @@ namespace LibSerial
          * @param charbuffer The character read from the serial port.
          * @param msTimeout The timeout period in milliseconds.
          */
-        void ReadByte(char&              charBuffer,
-                      const unsigned int msTimeout = 0);
+        void ReadByte(char&         charBuffer,
+                      const ssize_t msTimeout = 0);
 
         /**
          * @brief Reads a single byte from the serial port.
@@ -349,8 +349,8 @@ namespace LibSerial
          * @param charbuffer The character read from the serial port.
          * @param msTimeout The timeout period in milliseconds.
          */
-        void ReadByte(unsigned char&     charBuffer,
-                      const unsigned int msTimeout = 0);
+        void ReadByte(unsigned char& charBuffer,
+                      const ssize_t  msTimeout = 0);
 
         /**
          * @brief Reads a line of characters from the serial port.
@@ -366,17 +366,17 @@ namespace LibSerial
          * @param msTimeout The timeout value to return if a line termination
          *        character is not read.
          */
-        void ReadLine(std::string&       dataString,
-                      const char         lineTerminator = '\n',
-                      const unsigned int msTimeout = 0);
+        void ReadLine(std::string&   dataString,
+                      const char     lineTerminator = '\n',
+                      const ssize_t  msTimeout = 0);
 
         /**
          * @brief Writes a character array buffer to the serial port.
          * @param charBuffer The character array to be written to the serial port.
          * @param numberOfBytes The number of bytes to write to the serial port.
          */
-        void Write(const char*        charBuffer,
-                   const unsigned int numberOfBytes);
+        void Write(const char*   charBuffer,
+                   const ssize_t numberOfBytes);
 
         /**
          * @brief Writes a character array buffer to the serial port.
@@ -384,7 +384,7 @@ namespace LibSerial
          * @param numberOfBytes The number of bytes to write to the serial port.
          */
         void Write(const unsigned char* charBuffer,
-                   const unsigned int   numberOfBytes);
+                   const ssize_t        numberOfBytes);
 
         /**
          * @brief Writes a DataBuffer vector to the serial port.
